+++ perun/logic/runner.py
@@ -15,10 +15,10 @@ import perun.utils.decorators as decorators
 import perun.workload as workloads
 
 from perun.utils import get_module
-from perun.utils.structs import GeneratorSpec
+from perun.utils.structs import GeneratorSpec, Unit
 from perun.utils.helpers import COLLECT_PHASE_COLLECT, COLLECT_PHASE_POSTPROCESS, \
     COLLECT_PHASE_CMD, COLLECT_PHASE_WORKLOAD, CollectStatus, PostprocessStatus, \
-    Job, Unit
+    Job
 from perun.workload.singleton_generator import SingletonGenerator
 
 __author__ = 'Tomas Fiedor'